Added Tomography base class to simplify routines and reduce repeated code
