+++ lightworks/tomography/gate_fidelity.py
@@ -18,11 +18,11 @@ from numpy.typing import NDArray
 from lightworks.sdk.state import State
 
 from .mappings import PAULI_MAPPING, RHO_MAPPING
-from .process_tomography import ProcessTomography
+from .process_tomography import _ProcessTomography
 from .utils import _calculate_density_matrix, _combine_all, _vec
 
 
-class GateFidelity(ProcessTomography):
+class GateFidelity(_ProcessTomography):
     """
     Computes the average gate fidelity using equation 19 from
     https://arxiv.org/pdf/quant-ph/0205035, which involves performing state
@@ -33,8 +33,8 @@ class GateFidelity(ProcessTomography):
         n_qubits (int) : The number of qubits that will be used as part of the
             tomography.
 
-        base_circuit (PhotonicCircuit) : An initial circuit which produces the
-            required output state and can be modified for performing tomography.
+        base_circuit (PhotonicCircuit) : An initial circuit which realises the
+            required operation and can be modified for performing tomography.
             It is required that the number of circuit input modes equals 2 * the
             number of qubits.
