+++ lightworks/tomography/process_tomography.py
@@ -18,6 +18,7 @@ from lightworks.sdk.state import State
 
 from .experiments import ProcessTomographyExperiment, ProcessTomographyList
 from .mappings import INPUT_MAPPING, MEASUREMENT_MAPPING
+from .tomography import _Tomography
 from .utils import (
     TomographyDataError,
     _combine_all,
@@ -26,61 +27,13 @@ from .utils import (
 )
 
 
-class ProcessTomography:
+class _ProcessTomography(_Tomography):
     """
     Process tomography base class, implements some of the common methods
     required across different approaches.
-
-    Args:
-
-        n_qubits (int) : The number of qubits that will be used as part of the
-            tomography.
-
-        base_circuit (PhotonicCircuit) : An initial circuit which produces the
-            required output state and can be modified for performing tomography.
-            It is required that the number of circuit input modes equals 2 * the
-            number of qubits.
-
     """
 
     _tomo_inputs: tuple[str, ...] = ("Z+", "Z-", "X+", "Y+")
-
-    def __init__(
-        self,
-        n_qubits: int,
-        base_circuit: PhotonicCircuit,
-    ) -> None:
-        # Type check inputs
-        if not isinstance(n_qubits, int) or isinstance(n_qubits, bool):
-            raise TypeError("Number of qubits should be an integer.")
-        if not isinstance(base_circuit, PhotonicCircuit):
-            raise TypeError("Base circuit should be a circuit object.")
-
-        if 2 * n_qubits != base_circuit.input_modes:
-            msg = (
-                "Number of circuit input modes does not match the amount "
-                "required for the specified number of qubits, expected "
-                f"{2 * n_qubits}."
-            )
-            raise ValueError(msg)
-
-        self._n_qubits = n_qubits
-        self._base_circuit = base_circuit
-
-    @property
-    def base_circuit(self) -> PhotonicCircuit:
-        """
-        The base circuit which is to be modified as part of the tomography
-        calculations.
-        """
-        return self._base_circuit
-
-    @property
-    def n_qubits(self) -> int:
-        """
-        The number of qubits within the system.
-        """
-        return self._n_qubits
 
     def get_experiments(self) -> ProcessTomographyList:
         """
