+++ lightworks/tomography/tomography.py
@@ -1,0 +1,68 @@
+# Copyright 2024 Aegiq Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from lightworks.sdk.circuit import PhotonicCircuit
+
+
+class _Tomography:
+    """
+    Base class for all tomography experiments, which contains common logic
+    shared across them.
+
+    Args:
+
+        n_qubits (int) : The number of qubits that will be used as part of the
+            tomography.
+
+        base_circuit (PhotonicCircuit) : An initial circuit which can be
+            modified for performing tomography. It is required that the number
+            of circuit input modes equals 2 * the number of qubits.
+
+    """
+
+    def __init__(
+        self,
+        n_qubits: int,
+        base_circuit: PhotonicCircuit,
+    ) -> None:
+        # Type check inputs
+        if not isinstance(n_qubits, int) or isinstance(n_qubits, bool):
+            raise TypeError("Number of qubits should be an integer.")
+        if not isinstance(base_circuit, PhotonicCircuit):
+            raise TypeError("Base circuit should be a circuit object.")
+
+        if 2 * n_qubits != base_circuit.input_modes:
+            msg = (
+                "Number of circuit input modes does not match the amount "
+                "required for the specified number of qubits, expected "
+                f"{2 * n_qubits}."
+            )
+            raise ValueError(msg)
+
+        self._n_qubits = n_qubits
+        self._base_circuit = base_circuit
+
+    @property
+    def base_circuit(self) -> PhotonicCircuit:
+        """
+        The base circuit which the tomography experiment will be performed on.
+        """
+        return self._base_circuit
+
+    @property
+    def n_qubits(self) -> int:
+        """
+        The number of qubits within the system.
+        """
+        return self._n_qubits
